+++ src/app/features/pictures/pictures.module.ts
@@ -1,0 +1,19 @@
+import { NgModule } from '@angular/core';
+
+import { SharedModule } from '@shared/shared.module';
+
+import { PicturesRoutingModule } from './pictures-routing.module';
+import { PicturesComponent } from '@features/pictures/pictures.component';
+
+
+
+@NgModule({
+  declarations: [
+    PicturesComponent
+  ],
+  imports: [
+    SharedModule,
+    PicturesRoutingModule
+  ]
+})
+export class PicturesModule { }