+++ changelog.md
@@ -16,6 +16,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [0.2.3] - Changes after review with partners
 
+### Add
+
+- `feature` Create IPicture model, used for managing event images.
+
 ### Update
 
 - `style` Small sizes for social icons in mobile devices.
@@ -23,7 +27,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Fix
 
 - `feature` Fav icon enabled after login.
-
 
 ## [0.2.2] - Improvements in social features